% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{applyChangeMeta}
\alias{applyChangeMeta}
\title{Apply Meta Data Changes.}
\usage{
applyChangeMeta(GADSdat, changeTable)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{changeTable}{Change table.}
}
\value{
Returns the complete meta data sheet with additional columns for changes in meta data.
}
\description{
Function to apply meta data changes to a GADSdat object specified by a change table extracted by getChangeMeta.
}
\details{
The function ....
}
\examples{
# Example data set
to be done

}
