% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{applyLookup}
\alias{applyLookup}
\title{Recoade via lookup table.}
\usage{
applyLookup(GADSdat, lookup, suffix = "_r")
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{lookUp}{LookUp table created by \code{\link{createLookup}} and - if necessary -  collapsed by \code{\link{collapseColumns}}.}
}
\value{
Returns a recoded \code{GADSdat}.
}
\description{
Extract values form variables of \code{GADSdat} object and write to \code{xlsx} for manuel recoding.
}
\details{
If recoding of one or multiple variables is more complex, a lookup table can be created for later importing via \code{eatGADS}.
}
