% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_recode_multiMC.R
\name{applyLookup_expandVar}
\alias{applyLookup_expandVar}
\title{Recode via lookup table into multiple variables.}
\usage{
applyLookup_expandVar(GADSdat, lookup)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{lookup}{Lookup table created by \code{\link{createLookup}}.}
}
\value{
Returns a recoded \code{GADSdat}.
}
\description{
Recode one or multiple variables based on a lookup table created via \code{\link{createLookup}}. In contrast to \code{\link{applyLookup}},
this function allows the creation of multiple resulting variables from a single input variable.
}
\details{
tbd
}
\examples{
#to be done

}
