% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumCheck.R
\name{applyNumCheck}
\alias{applyNumCheck}
\title{Apply recodes according to a numerical check data.frame.}
\usage{
applyNumCheck(GADSdat, numCheck)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{numCheck}{A \code{data.frame} as created by \code{\link{createNumCheck}}.}
}
\value{
A recoded \code{GADSdat}.
}
\description{
Applies recodes as specified by a \code{numCheck} \code{data.frame}, as created by \code{\link{createNumCheck}}.
}
\details{
This function is currently under development.
}
\examples{
# tbd

}
