% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assimilateValLabels.R
\name{assimilateValLabels}
\alias{assimilateValLabels}
\title{Assimilate value labels.}
\usage{
assimilateValLabels(GADSdat, varNames, lookup = NULL)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{varNames}{Character string of a variable name.}

\item{lookup}{Look up \code{data.frame}.}
}
\value{
Returns the \code{GADSdat} object with changed meta data and recoded values.
}
\description{
Assimilate value labels of multiple variables as part of a \code{GADSdat} or \code{all_GADSdat} object.
}
\details{
Assimilation can be performed using all existing value labels or a look up table containing at least all existing value labels.
}
\examples{
# Change existing value labels
pisa2 <- changeValLabels(pisa, varName = "repeated",
                        value = c(1, 2),
                        valLabel = c("no grade repetition", "grade repitition"))

# Add value label to unlabeled value
mtcars_g <- import_DF(mtcars)
mtcars_g2 <- changeValLabels(mtcars_g, varName = "cyl",
                             value = c(4, 6, 8),
                             valLabel = c("four", "six", "eight"))

}
