% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoRecode.R
\name{autoRecode}
\alias{autoRecode}
\title{Auto-recode a variable in a \code{GADSdat}.}
\usage{
autoRecode(GADSdat, var, suffix = "", csv_path = NULL, template = NULL)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{var}{Character string of the variable name which should be sorted.}

\item{suffix}{Character string of the variable suffix.}

\item{csv_path}{Path for the \code{.csv} file for the look up table.}

\item{template}{Existing look up table.}
}
\value{
Returns a \code{GADSdat} object.
}
\description{
Auto-recode a variable in a \code{GADSdat}. A lookup table is created containing the respective recodes.
}
\examples{
# tbd

}
