% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.R
\name{cbind.GADSdat}
\alias{cbind.GADSdat}
\title{Bind two \code{GADSdat} objects into a single \code{GADSdat} object by columns.}
\usage{
\method{cbind}{GADSdat}(..., deparse.level = 1)
}
\arguments{
\item{...}{Multiple \code{GADSdat} objects imported via \code{eatGADS}.}

\item{deparse.level}{Argument is ignored in this method.}
}
\value{
Returns a \code{GADSdat} object.
}
\description{
Is a secure way to \code{cbind} the data and the meta data of two \code{GADSdat} objects. Currently, only limited merging options are supported.
}
\details{
If there are duplicate variables (except the variables specified in the \code{by} argument), these variables are removed from y.
The meta data is joined for the remaining variables via \code{rbind}.
}
