% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_metadata_handling.R
\name{changeVarNames}
\alias{changeVarNames}
\title{Change Variable Names.}
\usage{
changeVarNames(GADSdat, oldNames, newNames)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{oldNames}{Vector containing the old variable names.}

\item{newNames}{Vector containing the new variable names, in identical order as oldNames.}
}
\value{
Returns the GADSdat object with changed variable names.
}
\description{
Change variable names of a \code{GADSdat} or \code{all_GADSdat} object.
}
\details{
Applied to a \code{GADSdat} or \code{all_GADSdat} object, this function is a wrapper of \code{\link{getChangeMeta}} and \code{\link{applyChangeMeta}}
}
\examples{
# Example data set
#to be done

}
