% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4SPSS.R
\name{check4SPSS}
\alias{check4SPSS}
\title{Check \code{SPSS} Compliance of Meta Data}
\usage{
check4SPSS(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}
}
\value{
Returns a list with the entries \code{"varLabels"}, \code{"valLabels"} and \code{"missings"}.
}
\description{
Function to check if variable labels, value labels and missing codes comply with \code{SPSS} requirements for meta data.
}
\details{
The function measures the length of variable labels (\code{"varLabels"}, maximum of 256 characters),
value labels (\code{"valLabels"}, maximum of 120 characters) and
missing codes (\code{"missings"}, maximum of three missing codes for character variables)
of a \code{GADSdat} object and reports back violations on variable level.
}
\examples{
# Change example data set (create a value label with incorrect missing code)
pisa2 <- changeValLabels(pisa, varName = "computer_age",
                        value = 5, valLabel = "missing: No computer use")

pisa3 <- checkMissings(pisa2)

}
