% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4Stata.R
\name{check4Stata}
\alias{check4Stata}
\title{Check a \code{GADSdat} for compatibility with \code{Stata}.}
\usage{
check4Stata(GADSdat, version = c("Stata", "Stata 19/BE", "Stata 19/MP"))
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{version}{Optional single string to request checks for a specific Stata version
(see details).}
}
\value{
Either \code{NULL} if all checks are passed successfully, or a \code{list} of all
 check results (see details for explanations of the keywords) if any problem was detected.
}
\description{
This function performs all relevant checks to assess if a \code{GADSdat} complies with all of
 Stata's dataset requirements. Run this before exporting a dataset as \code{.dta}, using
 \code{\link{write_stata}}.
}
\details{
Specifically, the following requirements are tested:
\tabular{ll}{
 \code{dots_in_varNames}* \tab Variable names do not contain dots (\code{\link{checkVarNames}}). \cr
 \code{special_chars_in_varNames}* \tab Variable names do not contain special characters. \cr
 \code{varName_length}* \tab Variable names are not longer than the specific limit (\code{\link{checkVarNames}}). \cr
 \code{labeled_fractionals}* \tab There are no labeled fractional values (\code{\link{checkLabeledFractionals}}). \cr
 \code{large_integers}* \tab All labeled values can be coerced \code{as.integer} (\code{\link{checkIntOverflow}}). \cr
 \code{varLabel_length} \tab Variable labels are not longer than the specific limit (\code{\link{checkVarLabels}}). \cr
 \code{valLabel_length} \tab Value labels are not longer than the specific limit (\code{\link{checkValLabels}}). \cr
 \code{long_strings} \tab String variables do not contain string values that are longer than the specific limit. \cr
 \code{too_many_rows}* \tab The number of rows/observations does not exceed the specific limit. \cr
 \code{too_many_cols}* \tab The number of columns/variables does not exceed the specific limit. \cr
}
Not complying with the marked (*) requirements will prevent a dataset from being exported to a
 \code{.dta} file. Issues with unmarked requirements will be solved automatically by truncating.

Limits to different aspects of the dataset vary between versions of the software. By default
 (\code{version = "Stata"}), compliance with the limits for \code{Stata 19/SE} is checked.
 Checks against the limits for \code{Stata 19/BE} or \code{Stata 19/MP} can be requested by
 specifying \code{version} with the corresponding string. For more details, see \link{program_limits}.
}
\examples{
check4Stata(pisa)
}
\seealso{
Other dataset compliance checks: 
\code{\link{check4SPSS}()}
}
\concept{dataset compliance checks}
