% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIntOverflow.R
\name{checkIntOverflow}
\alias{checkIntOverflow}
\title{Check a \code{GADSdat} for large integers.}
\usage{
checkIntOverflow(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
Returns a \code{data.frame}, listing the \code{varName}s,
large \code{value}s (\code{as.numeric}), their \code{missings} tag
according to the metadata, and whether they actually occur in the data.
The \code{rownum}s of the affected rows in \code{GADSdat$labels} is provided
in a separate column as a fail safe.
}
\description{
Check a \code{GADSdat} object for any occurrences of integer values, esp.
in the metadata, that are too large for R to handle.
in its meta data which could cause problems esp. when exporting into \code{.dta} format.
}
\details{
According to its documentation, R can only handle \code{\link[base:integer]{integers}}
of up to \emph{roughly} \eqn{\pm 2 \times 10^9} (2,147,483,647 to be exact).
Right now, this restriction appears to only be of interfering relevance when
exporting a data set to \code{.dta} and only when the very large value is also
labeled (or tagged as missing). This is due to Stata only accepting integers
as being labeled - otherwise the value will stay a generic numeric.
}
\examples{
# Introduce a fractional value into meta data
pisa2 <- changeMissings(GADSdat = pisa,
                        varName = "schtype",
                        value = 9999999999,
                        missings = "miss")
checkIntOverflow(pisa2)

}
