% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLabeledFractionals.R
\name{checkLabeledFractionals}
\alias{checkLabeledFractionals}
\title{Check meta data for fractional values.}
\usage{
checkLabeledFractionals(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
Returns a list of two \code{data.frames}, respectively listing the
\code{varName}s and fractional \code{value}s tagged as \code{valid} or \code{miss}.
}
\description{
Check the a \code{GADSdat} object for any occurrences of fractional values
in its meta data which could cause problems esp. when exporting into \code{.dta} format.
}
\details{
This functions checks values that are tagged as being labeled in the corresponding
meta data column (\code{GADSdat$labels$labeled}). It therefore covers both
"truely" labeled values (that have been assigned a \code{valLabel}) and
values tagged as \code{missings} (with or without a \code{valLabel}).
}
\examples{
# Introduce a fractional value into meta data
pisa2 <- recodeGADS(GADSdat = pisa,
                    varName = "schtype",
                    oldValues = 2,
                    newValues = .5)
checkLabeledFractionals(pisa2)

}
