% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_trend.R
\name{checkLEStructure}
\alias{checkLEStructure}
\title{Checks compatability of GADS data bases with a linking error data base.}
\usage{
checkLEStructure(filePath1, filePath2, lePath)
}
\arguments{
\item{filePath1}{Path of the first GADS db file.}

\item{filePath2}{Path of the second GADS db file.}

\item{lePath}{Path of the linking error GADS db file.}
}
\value{
Returns a report list.
}
\description{
This function checks if a linking error data base is compatible with the two trend GADS data bases. For checking the compatability of two GADS data bases see \code{\link{checkTrendStructure}}.
}
\details{
This function inspects whether all linking error variables correspond to variables in the GADS data base and if the key variables also correspond to existing variables in the trend GADS data bases.
}
\examples{
# See vignette.

}
