% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{checkMissings}
\alias{checkMissings}
\title{Check and Adjust Missing Coding}
\usage{
checkMissings(GADSdat, missingLabel = "missing", addMissingCode = TRUE,
  addMissingLabel = TRUE)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{missingLabel}{Characatre how missing labels are named.}

\item{addMissingCode}{If [TRUE], missing codes are added according to occurence of [missingLabel] in [valLabel].}

\item{addMissingLabel}{If [TRUE], [generic missing] is added according to occurence of [mis] in [missings].}
}
\value{
Returns a GADSdat object
}
\description{
Function to check if missings are coded and labeled correctly in a GADSdat object.
}
\details{
The function compares value labels and missing codes of a GADSdat object and its labels data frame. Missmatches are reported and can be automatically adjusted.
}
\examples{
# Example data set
to be done

}
