% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_trend.R
\name{checkTrendStructure}
\alias{checkTrendStructure}
\title{Checks structure of trend gads and linking errors.}
\usage{
checkTrendStructure(all_GADSdat, pkList, fkList, filePath)
}
\arguments{
\item{filePath1}{Path of the first GADS db file.}

\item{filePath2}{Path of the second GADS db file.}

\item{lePath}{Path of the linking error db file. If NULL, no linking errors are added to the data.}
}
\value{
Create a data base.
}
\description{
tbd
}
\details{
tbd
}
\examples{
# See vignette.

}
