% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValLabels.R
\name{checkValLabels}
\alias{checkValLabels}
\title{Check value labels for length limits.}
\usage{
checkValLabels(
  GADSdat,
  vars = namesGADS(GADSdat),
  charLimits = c("SPSS", "Stata")
)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{Character vector with the variable names to which \code{checkValLabels()} should be applied.}

\item{classes}{Character vector with the classes to which \code{checkMissingLabels()} should be applied. Valid options are \code{"integer"}, \code{"double"}, and \code{"character"}.}

\item{valueRange}{[optional] Numeric vector of length 2: In which range should numeric values be checked?
If specified, only numeric values are returned and strings are omitted.}

\item{output}{Should the output be structured as a \code{"list"} or a \code{"data.frame"}?}
}
\value{
Returns a list of length \code{vars} or a \code{data.frame}.
}
\description{
Check if the value labels of a \code{GADSdat} comply with the length limits imposed by
 \code{SPSS} and/or \code{Stata}.
}
\details{
For details about program specific limits, see \link{getProgramLimit}.
}
\examples{
to be added

}
