% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVarNames.R
\name{checkVarNames}
\alias{checkVarNames}
\title{Check names for \code{SQLite} column name conventions and length limits.}
\usage{
checkVarNames(
  GADSdat,
  checkKeywords = TRUE,
  checkDots = TRUE,
  checkDuplicates = TRUE,
  charLimits = NULL
)
}
\arguments{
\item{GADSdat}{\code{GADSdat} or \code{all_GADSdat} object.}

\item{checkKeywords}{Logical. Should \code{SQLite} keywords be checked and modified?}

\item{checkDots}{Logical. Should occurrences of \code{"."} be checked and modified?}

\item{checkDuplicates}{Logical. Should case insensitive duplicate variable names be checked and modified?}

\item{charLimits}{Optional character vector of one or more program names(s) for the
limit check (see details). Currently, these are implemented: \code{c("SPSS", "Stata")}}
}
\value{
Returns the original object with updated variable names.
}
\description{
Checks names for \code{SQLite} column name conventions and \code{SPSS}/\code{Stata}
variable name limits, and applies appropriate variable name changes to
\code{GADSdat} or \code{all_GADSdat} objects.
}
\details{
Invalid column names in a \code{SQLite} data base include
\itemize{
\item \code{SQLite} keywords (see \code{\link[eatDB]{sqlite_keywords}}),
\item column names with a \code{"."} in it and
\item duplicate variable names which arise from \code{SQLite} being case insensitive.
}

The corresponding variable name changes are
\itemize{
\item appending the suffix \code{"Var"} to all \code{SQLite} keywords,
\item changing all \code{"."} in variable names to \code{"_"}, and
\item appending \code{"_2"} to case insensitive duplicated variable names.
}

Note that avoiding \code{"."} in variable names is beneficial for multiple reasons, such as
avoiding confusion with \code{S3} methods in \code{R} and issues when exporting to \code{Stata}.

\subsection{Variable name length limits}{
The length of variable names is limited to 64 \emph{bytes} in \code{SPSS} and to 32
 \emph{characters} in \code{Stata}. If more than one program name is provided in
 \code{charLimits}, the most restrictive among the chosen limits will be applied. Variable names
 exceeding that limit will be truncated and marked with the suffix \code{"_tr"}.
}
}
\examples{
# Change example data set (create an invalid variable name)
pisa2 <- changeVarNames(pisa, oldNames = "computer_age",
                        newNames = "computer.age")

pisa3 <- checkVarNames(pisa2)

}
