% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{clean_cache}
\alias{clean_cache}
\title{Clean temporary cache.}
\usage{
clean_cache(tempPath = tempdir())
}
\arguments{
\item{tempPath}{Local directory in which the data base was temporarily be stored.}
}
\value{
Returns nothing.
}
\description{
Deprecated. The cached data base is now cleaned when the R sessions ends automatically.
}
\details{
Cleans the temporary cache, speficied by tempdir(). This function should always be executed at the end of an \code{R} session if \code{\link{getGADS_fast}} or \code{\link{getTrendGADS}} with \code{fast = TRUE} has been used.
}
\examples{
\dontrun{
clean_cache()
}

}
