% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{collapseColumns}
\alias{collapseColumns}
\title{Collapse columns.}
\usage{
collapseColumns(lookup, recodeVars, prioritize)
}
\arguments{
\item{lookup}{A lookup table \code{data.frame} as created via \code{\link{createLookup}}.}

\item{recodeVars}{Character vector of variable names which should be collapsed.}

\item{prioritize}{Which of the variables should be prioritized, if multiple values are available?}
}
\value{
Returns a data frame that can be used for \code{\link{applyLookup}}.
}
\description{
Collapse columns of a lookup table created by \code{\link{createLookup}}.
}
\details{
tbd
}
\examples{
#to be done

}
