% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGADS.R
\name{compareGADS}
\alias{compareGADS}
\title{Compare two GADS.}
\usage{
compareGADS(GADSdat_old, GADSdat_new, varNames)
}
\arguments{
\item{GADSdat_old}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{GADSdat_new}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{varNames}{Character string of variable names to be compared.}
}
\value{
Returns a list with \code{"all equal"} or a \code{data.frame}.
}
\description{
Compare multiple variables of two \code{GADSdat} or \code{all_GADSdat} objects.
}
\details{
Returns \code{"all equal"} if the variable is identical across the objects or a \code{data.frame}
containing a frequency table with the values which have been changed. Especially useful for checks
after recoding.
}
\examples{
# Recode a GADS
pisa2 <- recodeGADS(pisa, varName = "schtype",
                        oldValues = 3, newValues = 9)

# Compare
compareGADS(pisa, pisa2, varNames = c("ganztag", "schtype"))

}
