% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_create_GADS.R
\name{createGADS}
\alias{createGADS}
\title{Create GADS data base.}
\usage{
createGADS(allList, pkList, fkList, filePath)
}
\arguments{
\item{allList}{An object created via mergeLabels.}

\item{pkList}{List of primary keys.}

\item{fkList}{List of foreign keys.}

\item{filePath}{Path to the db file to write (including name); has to end on '.db'.}
}
\value{
Creates a data base in the given path, returns NULL.
}
\description{
Creates a relational data base.
}
\details{
Uses createDB from the eatDB package to create a relational data base. For details on how to define keys see the documentation of createDB.
}
\examples{
# see createDB

}
