% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_trend.R
\name{createLEs}
\alias{createLEs}
\title{Create data base for Linking Errors.}
\usage{
createLEs(LEs, IDvars, filePath)
}
\arguments{
\item{LEs}{GADSdat object with linking errors.}

\item{IDvars}{Primary keys in the data frame.}

\item{filePath}{Path of the existing db file.}
}
\value{
Create a data base.
}
\description{
Create a data base that includes linking errors for trend reports.
}
\details{
Creates a single data table (with meta data) including the linking errors which are later used in \code{\link{getTrendGADS}}.
}
\examples{
# See vignette.

}
