% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumCheck.R
\name{createNumCheck}
\alias{createNumCheck}
\title{Create data.frame for specification of numerical plausibility checks.}
\usage{
createNumCheck(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
A data.frame with the following variables:
\item{variable}{All numerical variables in the \code{GADSdat}}
\item{varLabel}{Corresponding variable labels}
\item{min}{Minimum value for the specific variable.}
\item{max}{Maximum value for the specific variable.}
\item{value_new}{Which value should be inserted if values exceed the specified range?}
}
\description{
All numerical variables without value labels in a \code{GADSdat} are selected and a \code{data.frame} is created, which allows the specification
of minima and maxima.
}
\details{
This function is currently under development.
}
\examples{
# tbd

}
