% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyTheseVariables.R
\name{emptyTheseVariables}
\alias{emptyTheseVariables}
\title{Set variables to \code{NA}.}
\usage{
emptyTheseVariables(GADSdat, vars, label_suffix = "")
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{Character vector of variable names which should be set to \code{NA}.}

\item{label_suffix}{Suffix added to variable labels for the affected variables in the \code{GADSdat}.}
}
\value{
Returns the recoded \code{GADSdat}.
}
\description{
Set all values within one or multiple variables to \code{NA}.
}
\examples{
# empty multiple variables
pisa2 <- emptyTheseVariables(pisa, vars = c("idstud", "idschool"))
}
