% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_write_spss.R
\name{export_tibble}
\alias{export_tibble}
\title{Transform a GADSdat to a haven like tibble}
\usage{
export_tibble(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}
}
\value{
Returns a tibble.
}
\description{
Havens \code{\link[haven]{read_spss}} stores data together with meta data (e.g. value and variable labels) in a tibble with attributes on variable level. This function transforms a \code{GADSdat} object to such a tibble.
}
\details{
...
}
\examples{
# tbd

}
