% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_preparation.R
\name{extractGADSdat}
\alias{extractGADSdat}
\title{Extract single GADSdat from all_GADSdat}
\usage{
extractGADSdat(all_GADSdat, name)
}
\arguments{
\item{all_GADSdat}{\code{all_GADSdat} object, containing a table with the corresponding GADSdat name}

\item{name}{A character vector with length 1 with the name of the single GADSdat}
}
\value{
Returns an \code{GADSdat} object.
}
\description{
Function to extract a single \code{GADSdat} from an \code{all_GADSdat} object
}
\details{
The function createDB takes a list of data frames and a single data frame with metainformation as input. This function transforms lists from import_SPSS and import_RDS into a list of data frames (in the same order as inputted, which is then used for the merging order in createDB). Additionally the seperate lists of metainformation for each data frame are merged and a data frame unique identifier is added.
}
\examples{
# Example data set
#to be done

}
