% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixEncoding.R
\name{fixEncoding}
\alias{fixEncoding}
\title{Fix encoding problems.}
\usage{
fixEncoding(x, input = c("other", "ASCII"))
}
\arguments{
\item{x}{A character vector or \code{GADSdat} object.}

\item{input}{Which encoding was used in \code{\link{import_spss}}.}
}
\value{
The modified character vector or \code{GADSdat} object.
}
\description{
Fix some encoding problems of a character vector or a \code{GADSdat} object, which was encoded presumably using \code{UTF-8}
and imported using \code{ASCII} encoding.
}
\examples{
# tbd

}
