% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_program_limit.R
\name{get_program_limit}
\alias{get_program_limit}
\title{Get program specific limits}
\usage{
get_program_limit(
  program = c("SPSS", "Stata"),
  component = c("varname", "varlabel", "vallabel", "stringvar", "rows", "cols"),
  version = NULL
)
}
\arguments{
\item{program}{Character. Name(s) of the program(s) whose limits should be returned (see details).}

\item{component}{Single string. Which limits should be returned?}

\item{version}{Optional. Overwrites \code{program} with a specific version of a program (see details).}
}
\value{
A list of two elements: \code{x} (numeric size of the limit) and
 \code{unit} ("char", "byte", or "generic")
}
\description{
Get different limits for datasets specific to \code{SPSS} and \code{Stata}. This function
 primarily serves as a helper to the (sub)functions of the dataset compliance check family.
}
\details{
\code{SPSS} and \code{Stata} impose different limits to their data sets.
}
\examples{
get_program_limit("SPSS", "varname")
}
\references{
\href{https://www.stata.com/products/comparison-of-limits/}{Stata: Comparison of limits}
}
\seealso{
Other dataset compliance checks: 
\code{\link{checkIntOverflow}()},
\code{\link{checkLabeledFractionals}()}
}
\concept{dataset compliance checks}
