% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{getChangeMeta}
\alias{getChangeMeta}
\title{Extract table for Meta Data Changes.}
\usage{
getChangeMeta(GADSdat, changeCol = c("varName", "varLabel", "format",
  "display_width", "class", "value", "valLabel", "missings"))
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{changeCol}{Names of columns on which changes are to be specified.}
}
\value{
Returns the complete meta data sheet with additional columns for changes in meta data.
}
\description{
Function to obtain a data frame from a GADSdat object for meta data changes. Changes to value are currently not supported.
}
\details{
The function ....
}
\examples{
# Example data set
to be done

}
