% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_eatDB_UI.R
\name{getGADS}
\alias{getGADS}
\title{Get data from GADS data base.}
\usage{
getGADS(vSelect = NULL, filePath)
}
\arguments{
\item{vSelect}{Variables}

\item{filePath}{Path of the existing db file.}
}
\value{
Returns a long format data frame including variable names, labels, values, value labels and missing labels.
}
\description{
Extracts variables from a GADS data base.
}
\details{
See createDB and dbPull for further explanation of the query and merging processes.
}
\examples{
# See vignette.

}
