% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{getGADS_fast}
\alias{getGADS_fast}
\title{Get data from GADS data base fast from server directory.}
\usage{
getGADS_fast(vSelect = NULL, filePath, tempPath)
}
\arguments{
\item{vSelect}{Variables}

\item{filePath}{Path of the existing db file.}

\item{tempPath}{Local directory in which the dataBase can temporarily be stored.}
}
\value{
Returns a GADSdat object.
}
\description{
Extracts variables from a GADS data base. Uses a local temporary directory to speed up loading the GADS from a server.
}
\details{
See createDB and dbPull for further explanation of the query and merging processes.
}
\examples{
# See vignette.

}
