% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProgramLimit.R
\name{getProgramLimit}
\alias{getProgramLimit}
\title{Get program specific limits}
\usage{
getProgramLimit(
  program = c("SPSS", "Stata", "Stata 19/BE", "Stata 19/MP"),
  component = c("varNames", "varLabels", "valLabels", "stringvars", "nrows", "ncols"),
  version = NULL
)
}
\arguments{
\item{program}{Character vector of the programs/program version that should be considered.}

\item{component}{Single string. Which limits should be returned?}
}
\value{
A list of two elements: \code{x} (numeric size of the limit) and
 \code{unit} ("char", "byte", or "generic"). If more than one \code{program} is specified and
 \code{component} is not "all", the most restrictive limit will be returned.
}
\description{
Get the (most restrictive) limit that \code{SPSS} and/or \code{Stata} imposes on a specific
 aspect of a data set.
}
\details{
\code{SPSS} and \code{Stata} impose different limits to different components of their data sets,
 e.g., on the length of variable names. For details and a list of relevant limits, see \link{program_limits}.
 Additionally, limits may vary between software versions. This primarily applies to \code{Stata}'s
 product tiers, but also for (very) old \code{SPSS} version.

For \code{program}, \code{"SPSS"} and \code{"Stata"} imply \code{SPSS 30} and \code{Stata 19/SE},
 respectively, as these are the most relevant version among the implemented ones.

If more than one
}
\examples{
# Get all limits that Stata imposes on its datasets
eatGADS:::get_program_limit("Stata", "all")

# Get the specific limit on variable name lengths under SPSS
eatGADS:::get_program_limit("SPSS", "varNames")
}
\seealso{
Other dataset compliance checks: 
\code{\link{checkIntOverflow}()},
\code{\link{checkLabeledFractionals}()}
}
\concept{dataset compliance checks}
