% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_trend.R
\name{getTrendGADS}
\alias{getTrendGADS}
\title{Get data for trend reports.}
\usage{
getTrendGADS(vSelect = NULL, filePath1, filePath2, lePath = NULL,
  years)
}
\arguments{
\item{vSelect}{Variables}

\item{filePath1}{Path of the first GADS db file.}

\item{filePath2}{Path of the second GADS db file.}

\item{lePath}{Path of the linking error db file. If NULL, no linking errors are added to the data.}

\item{years}{A numeric vector of length 2. The first year corresponds to filePath1, the second year to filePath2.}
}
\value{
Returns a GADSdat object.
}
\description{
Extracts variables from a two GADS data base and a linking error data base.
}
\details{
This function extracts data from two GADS data bases and a linking error data base. The data is merged and can further be used via \code{\link{extractData}}. See createDB and dbPull for further explanation of the query and merging processes.
}
\examples{
# See vignette.

}
