% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrend3GADS.R
\name{getTrendsGADS}
\alias{getTrendsGADS}
\title{Get data for trend reports.}
\usage{
getTrendsGADS(
  filePaths,
  lePath = NULL,
  vSelect = NULL,
  years,
  fast = TRUE,
  tempPath = tempdir()
)
}
\arguments{
\item{filePaths}{Character vectors with paths to the \code{eatGADS} db files.}

\item{lePath}{Path of the linking error db file. If \code{NULL}, no linking errors are added to the data.}

\item{vSelect}{Variables from all GADS to be selected (as character vector).}

\item{years}{A numeric vector with identical length as \code{filePaths}.}

\item{fast}{Should \code{\link{getGADS_fast}} be used for data loading instead of \code{\link{getGADS}}? Using the default is heavily recommended.}

\item{tempPath}{The directory, in which both GADS will be temporarily stored. Using the default is heavily recommended.}
}
\value{
Returns a \code{GADSdat} object.
}
\description{
Extracts variables from multiple \code{eatGADS} data bases and a linking error data base.
Data can then be extracted from the \code{GADSdat} object via
\code{\link{extractData}}. For extracting meta data from a data base or a \code{GADSdat} object see \code{\link{extractMeta}}. To speed
up the data loading, \code{\link{getGADS_fast}} is used per default.
}
\details{
This function extracts data from multiple GADS data bases and a linking error data base. All data bases have to be created via
\code{\link{createGADS}}. The two GADS are joined via \code{rbind()} and a variable \code{year} is added, corresponding to the
argument \code{years}. If \code{lePath} is specified, linking errors are also extracted and then merged to the GADS data. Make
sure to also extract the key variables necessary for merging the linking errors (the domain variable for all linking errors,
additionally the competence level variable for linking errors for competence levels). The \code{GADSdat} object can then further
be used via \code{\link{extractData}}. See \code{\link[eatDB]{createDB}} and \code{\link[eatDB]{dbPull}} for further explanation
of the querying and merging processes.
}
\examples{
# See getGADS vignette


}
