% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_loading.R
\name{import_DF}
\alias{import_DF}
\title{Import R data frame}
\usage{
import_DF(df, checkVarNames = TRUE)
}
\arguments{
\item{filePath}{Source file location.}
}
\value{
Returns a list with a) the actual data and b) a data frame with all variable and value labels in long format.
}
\description{
Function to import a data frame from R.
}
\details{
...
}
\examples{
# Example data set
to be done

}
