% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_loading.R
\name{import_raw}
\alias{import_raw}
\title{Import R data frame with explicit meta data sheets}
\usage{
import_raw(df, varLabels, valLabels, checkVarNames = TRUE)
}
\arguments{
\item{df}{A data frame (without factors).}

\item{varLabels}{A data frame containing the variable labels. All variables in the data have to have exactly one column in this data.frame.}

\item{valLabels}{A data frame containing the value labels. All referenced variables have to appear in the data, but not all variables in the data have to receive value labels.}

\item{checkVarNames}{Should variable names be checked for vioalitions of SQLite and R naming rules?}
}
\value{
Returns a list with the actual data \code{dat} and a data frame with all meta information in long format \code{labels}.
}
\description{
Function to import a \code{data.frame} object for use in \code{eatGADS} while adding explicit variable and value meta information through separate \code{data.frame}.
}
\details{
The varLables \code{data.frame} has to contain exactly two variables, namely \code{varName} and \code{varLabel}. The valLables \code{data.frame} has to contain exactly four variables, namely \code{varName}, \code{value}, \code{valLabel} and \code{missings}. The column \code{value} can only contain numerical values. The column \code{missings} can only contain the values \code{"valid"} and \code{"miss"}.
}
