% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_loading.R
\name{import_RDS}
\alias{import_RDS}
\title{Import R data}
\usage{
import_RDS(filePath)
}
\arguments{
\item{filePath}{Source file location.}
}
\value{
Returns a list with a) the actual data and b) a data frame with all variable and value labels in long format.
}
\description{
Function to import an RDS-saved file (maybe change to list??)
}
\details{
...
}
\examples{
# Example data set
to be done

}
