% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_loading.R
\name{import_spss}
\alias{import_spss}
\title{Import SPSS data}
\usage{
import_spss(filePath, checkVarNames = TRUE, labeledStrings = FALSE)
}
\arguments{
\item{filePath}{Source file location, ending on .sav}
}
\value{
Returns a list with a) the actual data and b) a data frame with all variable and value labels in long format.
}
\description{
Function to import sav files while extracting variable and value labels.
}
\details{
SPSS files (.sav) store variable and value labels. Via import_spss data frames are imported and transformed to data frames. The meta-information is stored seperately in a long format data frame. Important Note: To get labels, missings have to be given explicit labels in SPSS! Additional missing column is generated.
}
\examples{
# Example data set
to be done

}
