% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertVariable.R
\name{insertVariable}
\alias{insertVariable}
\title{Order the variables in a \code{GADSdat}.}
\usage{
insertVariable(GADSdat, var, varBefore)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{var}{Character string of the variable name which should be sorted.}

\item{varBefore}{Character string of the variable name which should be followed by \code{var}.}
}
\value{
Returns a \code{GADSdat} object.
}
\description{
Order the variables in a \code{GADSdat} according to a character vector. If there are discrepancies between the two sets, a warning is issued.
}
\details{
The variables in the \code{dat} and in the \code{labels} section are ordered. Variables not contained in the character vector are moved to the end of the data.
}
