% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{labelsGADS}
\alias{labelsGADS}
\title{Labels from relational db.}
\usage{
labelsGADS(filePath)
}
\arguments{
\item{filePath}{Path of the existing db file.}
}
\value{
Returns a long format data frame including variable names, labels, values, value labels and missing labels.
}
\description{
Returns the variable and value labels of all variables in the GADS
}
\details{
Variable, value and missing labels as stored in the original SPSS-files and factors from R files are converted to long format for storage in the data base. labelsGADS returns them as a long format data frame.
}
\examples{
\dontrun{
metaData <- labelsGADS("t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")
View(metaData)
}

}
