% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{merge.GADSdat}
\alias{merge.GADSdat}
\title{Merge two GADSdat objects into a single GADSdat object.}
\usage{
\method{merge}{GADSdat}(x, y, by, all = TRUE, all.x = all, all.y = all, ...)
}
\arguments{
\item{x}{\code{GADSdat} object imported via eatGADS.}

\item{y}{\code{GADSdat} object imported via eatGADS.}

\item{by}{A character vector.}

\item{all}{A character vector (either a full join or an inner join).}

\item{all.x}{See merge.}

\item{all.y}{See merge.}

\item{...}{Further arguments are currently not supported but have to be included for R CMD checks.}
}
\value{
Returns a GADSdat object.
}
\description{
Is a secure way to merge the data and the meta data of two GADSdat objects. Only a very specific way of merging is possible.
}
\details{
If there are variables duplicate (the variables specified in by are excempt), these variables are removed from y. The data is merged via a full join. The meta data is joined for the remaining variables via rbind.
}
\examples{
# Example data set
#to be done

}
