% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_preparation.R
\name{mergeLabels}
\alias{mergeLabels}
\title{Prepare data and metadata}
\usage{
mergeLabels(...)
}
\arguments{
\item{...}{Lists including data and label element.}
}
\value{
Returns a list with a list of all data frames and a single data frame containing all variable and value labels.
}
\description{
Function to transform list of data-frames and metainformation ready for data base creation
}
\details{
The function createDB takes a list of data frames and a single data frame with metainformation as input. This function transforms lists from import_SPSS and import_RDS into a list of data frames (in the same order as inputted, which is then used for the merging order in createDB). Additionally the seperate lists of metainformation for each data frame are merged and a data frame unique identifier is added.
}
\examples{
# Example data set
to be done

}
