% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{miss2NA}
\alias{miss2NA}
\title{Recode Missings to NA}
\usage{
miss2NA(labeledDat)
}
\arguments{
\item{importedSPSS}{Data frame imported via [import_spss].}
}
\value{
Returns the data frame with NA instead of missing codes.
}
\description{
Recode Missings to NA according to missing labels in label data frame.
}
\details{
Missings are imported as their values via import_spss. Using the value labels in the labels data frame, miss2NA recodes these missings codes to NA.
}
\examples{
# Example data set
to be done

}
