% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesGADS.R
\name{namesGADS}
\alias{namesGADS}
\title{Variables names of a GADS.}
\usage{
namesGADS(GADS)
}
\arguments{
\item{GADS}{A \code{GADSdat} object, a \code{all_GADSdat} or the path to an existing \code{eatGADS} data base.}
}
\value{
Returns a character vector or a named list of character vectors.
}
\description{
Variables names of a \code{GADSdat} object, a \code{all_GADSdat} object or a \code{eatGADS} data base.
}
\details{
If the function is applied to a \code{GADSdat} object, a character vector with all variable names is returned. If the function is
applied to a \code{all_GADSdat} object or to the path of a \code{eatGADS} data base, a named list is returned. Each list entry
represents a data table in the object.
}
\examples{
\dontrun{
varNames <- namesGADS("t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")

# all variable names sorted by data table
varNames

# variables in a specific data table
varNames$allDat

}

}
