% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{orderLike}
\alias{orderLike}
\title{Order the variables a GADSdat.}
\usage{
orderLike(GADSdat, newOrder)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via eatGADS.}

\item{newOrder}{A character vector containing the order of variables.}
}
\value{
Returns a GADSdat object.
}
\description{
Order the variables in a GADSdat according to a character vector. If there are discrepancies between the two sets, a warning is issued.
}
\details{
The variables in the \code{dat} and in the \code{labels} section are ordered. Variables no contained in the character vector are moved to the end of the data.
}
\examples{
# Example data set
#to be done

}
