% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa.R
\docType{data}
\name{pisa}
\alias{pisa}
\title{PISA Plus Example Data}
\format{
A data.frame with 500 rows and 139 variables, including:
\describe{
  \item{idstud}{Person ID variable}
  \item{idschool}{School ID variable}
  \item{schtype}{School type}
  \item{...}{}
}
}
\source{
\url{https://www.iqb.hu-berlin.de/fdz/Datenzugang/CF-Antrag}
}
\usage{
pisa
}
\description{
A small example data set from the German PISA Plus campus files as distributed by the \code{Forschungsdatenzentrum, IQB}.
}
\keyword{datasets}
