% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeString2NA.R
\name{recodeString2NA}
\alias{recodeString2NA}
\title{Recode a string to \code{NA}.}
\usage{
recodeString2NA(GADSdat, recodeVars = namesGADS(GADSdat), string = "")
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{recodeVars}{Character vector of variable names which should be recoded.}

\item{string}{Which string should be recoded to \code{NA}?}
}
\value{
Returns the recoded \code{GADSdat}.
}
\description{
Recode an unlabeled string in multiple variables in a \code{GADSdat} to \code{NA}.
}
\details{
A check is performed, whether there are truly no labels given to the specified string. Number of recodes per variable are reported.
}
\examples{
# create example GADS
dat <- data.frame(ID = 1:4, var1 = c("", "Eng", "Aus", "Aus2"),
                  var2 = c("", "French", "Ger", "Ita"),
                  stringsAsFactors = TRUE)
gads <- import_DF(dat)

# recode empty strings
gads2 <- recodeString2NA(gads)

}
