% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_recode_multiMC.R
\name{remove_2NA_char}
\alias{remove_2NA_char}
\title{Shorten multiple text variables while giving NA codes.}
\usage{
remove_2NA_char(GADSdat, vars, max_num = 2, na_value)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{A character vector with the names of the text variables.}

\item{max_num}{Maximum number of text variables. Additional text variables will be removed and NA codes given accordingly.}

\item{na_value}{Which NA value should be given in cases of too many values on text variables.}
}
\value{
Returns the modified \code{GADSdat}.
}
\description{
Remove text variables from a certain number from \code{GADSdat} while coding overflowing answers as complete missings.
}
\details{
In some cases, multiple text variables contain the information of one variable (e.g. multiple answers to an open item).
}
\examples{
#to be done

}
