% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyValLabels.R
\name{removeEmptyValLabels}
\alias{removeEmptyValLabels}
\title{Remove unused value labels.}
\usage{
removeEmptyValLabels(GADSdat, vars, whichValLabels = c("miss", "valid", "all"))
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{vars}{Character string of variable names.}

\item{whichValLabels}{Should unused missing value tags and labels (\code{"miss"}).}
}
\value{
Returns the \code{GADSdat} object with changed meta data.
}
\description{
Remove unused value labels of a variable as part of a \code{GADSdat} object.
}
\examples{
#tbd

}
