% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{removeVars}
\alias{removeVars}
\title{Remove variables from a GADSdat.}
\usage{
removeVars(GADSdat, vars)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object.}

\item{vars}{A character vector containing the variables to be removed.}
}
\value{
Returns a GADSdat object.
}
\description{
Remove variables and their meta data from a \code{GADSdat} object.
}
\details{
Wraps removing the variable from the data.frame in the \code{GADSdat} object and \code{\link{updateMeta}}.
}
\examples{
# Example data set
#to be done

}
