% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_preparation.R
\name{splitGADS}
\alias{splitGADS}
\title{Split GADS}
\usage{
splitGADS(GADSdat, nameList)
}
\arguments{
\item{GADSdat}{\code{GADSdat} objects, as named arguments in the correct merge order.}

\item{nameList}{A list with character vector. The names in the list correspond the the hierarchy levels.}
}
\value{
Returns an \code{all_GADSdat} object, which consists of list with a list of all data frames \code{"datList"} and a single data frame containing all meta data information \code{"allLabels"}. For more details see also \code{\link{mergeLabels}}.
}
\description{
Function to split a \code{GADSdat} with all variables in it into hierarchical levels.
}
\details{
The function takes a \code{GADSdat} object and splits it into its desired hierarchical levels. Meta data include therefore then references via \code{data_table} to the corresponding hierarchical level.
}
\examples{
# Example data set
#to be done

}
