% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringSplit.R
\name{stringSplit}
\alias{stringSplit}
\title{Split a character variable.}
\usage{
stringSplit(GADSdat, varName, pattern)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{varName}{Character string of a variable name.}

\item{pattern}{Character string of the splitting pattern.}
}
\value{
Returns the \code{GADSdat} object with with the changed variable.
}
\description{
Split a character variable which is part of a \code{GADSdat} object into multiple character variables.
}
