% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_write_spss.R
\name{write_spss}
\alias{write_spss}
\title{Write a \code{GADSdat} object to \code{sav}}
\usage{
write_spss(GADSdat, filePath)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{filePath}{Path of \code{sav} file to write.}
}
\value{
Writes \code{sav} file to disc, returns \code{NULL}.
}
\description{
Write a \code{GADSdat} object, which contains meta information as value and variable labels to an SPSS file (\code{sav}).
}
\details{
Careful: This function is still slightly experimental.
}
\examples{
# tbd

}
