% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{applyChangeMeta}
\alias{applyChangeMeta}
\title{Apply Meta Data Changes.}
\usage{
applyChangeMeta(GADSdat, changeTable)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{changeTable}{Change table as provided by \code{\link{getChangeMeta}}.}
}
\value{
Returns the modified \code{GADSdat} object.
}
\description{
Function to apply meta data changes to a \code{GADSdat} object specified by a change table extracted by \code{\link{getChangeMeta}}.
}
\details{
Values for which the change columns contain \code{NA} remain unchanged. Changes to values are currently not supported.
}
\examples{
# Example data set
to be done

}
