% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{applyLookup}
\alias{applyLookup}
\title{Recoade via lookup table.}
\usage{
applyLookup(GADSdat, lookup, suffix = NULL)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{lookup}{Lookup table created by \code{\link{createLookup}} and - if necessary -  collapsed by \code{\link{collapseColumns}}. Column names should be \code{c("variable", "value", "value_new")}.}

\item{suffix}{Suffix to add to the existing variable names. If \code{NULL} the old variables will be overwritten.}
}
\value{
Returns a recoded \code{GADSdat}.
}
\description{
Recode one or multiple variables based on a lookup table created via \code{\link{createLookup}}.
}
\details{
tbd
}
\examples{
#to be done

}
