% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_metadata_handling.R
\name{changeVarLabels}
\alias{changeVarLabels}
\title{Change the variable label.}
\usage{
changeVarLabels(GADSdat, varName, varLabel)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{varName}{Character string of variable names.}

\item{varLabel}{Character string of the new variable labels.}
}
\value{
Returns the GADSdat object with changed meta data..
}
\description{
Change the variable label of a variable as part of a \code{GADSdat} or \code{all_GADSdat} object.
}
\details{
Applied to a \code{GADSdat} or \code{all_GADSdat} object, this function is a wrapper of \code{\link{getChangeMeta}} and \code{\link{applyChangeMeta}}
}
\examples{
# Example data set
#to be done

}
