% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4Stata.R
\name{check4Stata}
\alias{check4Stata}
\title{Check a \code{GADSdat} for compatibility with \code{Stata}.}
\usage{
check4Stata(GADSdat, version = c("Stata", "Stata 19/BE", "Stata 19/MP"))
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{version}{Optional single string to request checks for a specific Stata version
(see details).}
}
\value{
Either \code{NULL} if all checks are passed successfully, or a \code{list} of all
 check results (see details for explanations for the keywords) if any problem was detected,
 including a verdict whether any of the problems is critical (\code{"hard issue"}) or will
 be solved automatically in the export process by truncating (\code{"soft issue"}).
}
\description{
This function performs all relevant checks to assess if a \code{GADSdat} complies with all of
 Stata's dataset requirements. Run this before exporting a dataset as \code{.dta}, using
 \link{write_stata}.
}
\details{
Specifically, the following requirements are tested:
\tabular{rll}{
 dots_in_varNames + special_chars_in_varNames \tab - \tab Variable names do not contain dots or special characters. \cr
 varName_length \tab - \tab Variable names are not longer than the specific limit (\link{checkVarNames}). \cr
 labeled_fractionals \tab - \tab There are no labeled fractional values like \code{99.9} (\link{checkLabeledFractionals}). \cr
 large_integers \tab - \tab All labeled values can be coerced \code{as.integer} (\link{checkIntOverflow}). \cr
 varLabel_length + valLabel_length \tab - \tab Variable labels and value labels are not longer than the specific limits (\link{checkVarLabels}; \link{checkValLabels}). \cr
 long_strings \tab - \tab String variables do not contain string values that are longer than the specific limit. \cr
 too_many_rows + too_many_cols \tab - \tab The numbers of rows/observations and columns/variables do not exceed the specific limits. \cr
}
Limits to different aspects of the dataset differ between the versions of the software. By
 default (\code{version = "Stata"}), compliance with the limits for \code{Stata 19/SE} is
 checked. Checks against the limits for \code{Stata 19/BE} or \code{Stata 19/MP} can be
 requested by specifying \code{version} with the corresponding string.
 For more details, see \link{program_limits}.
}
\examples{
check4Stata(pisa)

}
\seealso{
Other dataset compliance checks: 
\code{\link{check4SPSS}()},
\code{\link{checkValLabels}()}
}
\concept{dataset compliance checks}
