% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4Stata.R
\name{check4Stata}
\alias{check4Stata}
\title{Check a \code{GADSdat} for compatibility with \code{Stata}.}
\usage{
check4Stata(GADSdat, version = NULL)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{version}{Optional single string to request checks for a different Stata version
(see details)}
}
\value{
A list of check results (enumerated corresponding to the list above), headed by an
 overall \code{verdict} if the \code{GADSdat} is compatible (\code{TRUE}) or if problems have
 been detected (\code{FALSE}).
}
\description{
This wrapper performs all relevant checks tp assess if a \code{GADSdat} complies with all of
 Stata's dataset requirements. Run this before exporting a dataset as \code{.dta}, using
 \link{write_stata}.
}
\details{
Specifically, the following requirements are tested:
\enumerate{
 \item Variable names do not contain dots or special characters.
 \item Variable names are not longer than the specific limit (\link{checkVarNames}).
 \item There are no labeled fractional values like \code{99.9} (\link{checkLabeledFractionals}).
 \item All labeled values can be coerced \code{as.integer} (\link{checkIntOverflow}).
 \item Variable labels and value labels are not longer than the specific limits
  (\link{checkVarLabels}; \link{checkValLabels}).
 \item String variables do not have value labels. Currently, this is not allowed for
   \code{GADSdat} objects, anyway.
 \item String variables do not contain string values that are longer than the specific limit.
 \item The numbers of rows/observations and columns/variables do not exceed the specific limits.
}
Limits to different aspects of the dataset differ between the versions of the software. By
 default, compliance with the limits for \code{Stata 19/SE} is checked. Checks against the
 limits for \code{Stata 19/BE} or \code{Stata 19/MP} can be requested by specifying
 \code{version} with the corresponding string. For more details, see \link{program_limits}.
}
\examples{
check4Stata(pisa)

}
\seealso{
Other dataset compliance checks: 
\code{\link{check4SPSS}()},
\code{\link{checkValLabels}()}
}
\concept{dataset compliance checks}
