% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLabeledFractionals.R
\name{checkLabeledFractionals}
\alias{checkLabeledFractionals}
\title{Check a \code{GADSdat} for labeled fractional values.}
\usage{
checkLabeledFractionals(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
Returns a \code{data.frame}, listing the affected \code{varName}s,
the labeled fractional \code{value}s, their respective \code{missings} tags,
and whether they actually occur in the data (\code{empty}).
}
\description{
Check a \code{GADSdat} object for any occurrences of fractional values in its metadata,
including both "truly" labeled values and values tagged as \code{missings}.
}
\details{
This function is mainly useful to ensure a data set can be saved as a \code{.dta} file.
Unlike, for example, SPSS, Stata only allows for integer values
(and so-called extended missing values) to be labeled
(\href{https://www.stata.com/manuals/u12.pdf#u12.6.3Valuelabels}{Stata manual: 12.6.3}).
Trying to export (meta) data with labeled fractional values would therefore cause problems
and run into an error from \code{haven}'s \code{\link[haven:write_dta]{write_dta}} function.
}
\examples{
# Introduce a fractional value into meta data
pisa2 <- recodeGADS(GADSdat = pisa,
                    varName = "schtype",
                    oldValues = 2,
                    newValues = .5)
checkLabeledFractionals(pisa2)

}
\seealso{
Other dataset compliance checks: 
\code{\link{checkIntOverflow}()}
}
\concept{dataset compliance checks}
