% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{checkMissings}
\alias{checkMissings}
\title{Check and Adjust Missing Coding}
\usage{
checkMissings(GADSdat, missingLabel = "missing", addMissingCode = TRUE,
  addMissingLabel = FALSE)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{missingLabel}{Single string indicating how missing labels are commonly named in the value labels.}

\item{addMissingCode}{If \code{TRUE}, missing codes are added according to occurence of \code{"missingLabel"} in \code{"valLabel"}.}

\item{addMissingLabel}{If \code{TRUE}, \code{"generic missing"} is added according to occurence of \code{"mis"} in \code{"missings"}. As often various value labels for missings are used, this argument should be used with great care.}
}
\value{
Returns a GADSdat object.
}
\description{
Function to check if missings are coded and labeled correctly in a GADSdat object.
}
\details{
The function compares value labels \code{"valLabels"} and missing codes \code{"missings"} of a \code{GADSdat} object and its meta data information. Missmatches are reported and can be automatically adjusted.
}
\examples{
# Example data set
#to be done

}
