% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValLabels.R
\name{checkValLabels}
\alias{checkValLabels}
\title{Check Value Labels}
\usage{
checkValLabels(GADSdat, vars = namesGADS(GADSdat))
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{Character vector with the variable names to which \code{checkValLabels()} should be applied.}
}
\value{
Returns a list of length 2. Each contains a list of the length of \code{vars}.
\item{labels with no values}{Value labels with no occurrence in the data}
\item{not labeled values}{Values in the data with no value labels}
}
\description{
Check value labels for (a) value labels with no occurrence in the data and (b) values with no value labels.
}
\details{
\code{NAs} are excluded from this check. Designated missing codes are reported normally.
}
\examples{
# Check a categorical and a metric variable
checkValLabels(pisa, vars = c("g8g9", "age"))

}
