% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValLabels.R, R/checkVarLabels.R
\name{checkValLabels}
\alias{checkValLabels}
\alias{checkVarLabels}
\title{Check Lengths of Labels.}
\usage{
checkValLabels(
  GADSdat,
  charLimits = c("SPSS", "Stata"),
  vars = namesGADS(GADSdat),
  printLength = 40
)

checkVarLabels(
  GADSdat,
  charLimits = c("SPSS", "Stata"),
  vars = namesGADS(GADSdat),
  printLength = 40
)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{charLimits}{Character vector of the program(s) against whose limit(s) the labels should
be checked.}

\item{vars}{Optional character vector of the variables whose value labels should be checked.
By default, all value labels will be checked.}

\item{printLength}{Single numeric value. The first n = \code{printLength} characters of each
long label will be reported. Alternatively, specify \code{NULL} to deactivate the truncation
and have the function report the full labels (not recommended; also see details).}
}
\value{
A \code{data.frame}, reporting every (truncated) long \code{varLabel}/\code{valLabel},
 their respective \code{length} in the relevant \code{unit} and the \code{varName} in which
 they occur. For \code{checkValLabels}, the labeled \code{value}, as well as whether that value
 actually occurs in the data (\code{empty}), is also reported.
}
\description{
Check if the value or variable labels of a \code{GADSdat} comply with the length limits imposed
 by \code{SPSS} or \code{Stata}.
}
\details{
If more than one program name is given in \code{charLimits}, the most restrictive limit will be
 applied. For details about program-specific limits, see \link{program_limits}.

Please note that setting \code{printLength} to \code{NULL} (and thereby deactivating label
 truncation) might not actually result in the printing of the full length of the exceedingly
 long labels if you are using RStudio. The program's own limits on the number of characters
 printed to the console may still apply
 (see \href{https://stackoverflow.com/questions/36800475/avoid-string-printed-to-console-getting-truncated-in-rstudio}{Stack Overflow}).
}
\section{Functions}{
\itemize{
\item \code{checkValLabels()}: Check value labels for length limits.

\item \code{checkVarLabels()}: Check variable labels for length limits.

}}
\examples{
# check value labels
pisa2 <- pisa
pisa2$labels[4, "valLabel"] <- paste0(rep("abcdefg", 4300), collapse = "")
eatGADS:::checkValLabels(pisa2)

# check variable labels
pisa2$labels[1, "varLabel"] <- paste0(rep("abcdefg", 12), collapse = "")
eatGADS:::checkVarLabels(pisa2)
}
