% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValue.R
\name{checkValue}
\alias{checkValue}
\title{Check for a specific value}
\usage{
checkValue(GADSdat, value)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{value}{Single string indicating how missing labels are commonly named in the value labels.}
}
\value{
Returns \code{NULL}.
}
\description{
Function to look for occurrences of a specific value in a \code{GADSdat}.
}
\details{
The function checks occurrences of a specific value in all variables in the \code{GADSdat} and outputs a message
containing a list of variables in which the value occurs.
}
\examples{
# Example data set
#to be done

}
