% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_cleaning.R
\name{checkVarNames}
\alias{checkVarNames}
\title{Check names for SQLite conventions.}
\usage{
checkVarNames(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} or \code{all_GADSdat} object imported via eatGADS.}
}
\value{
Returns the original object with updated variable names.
}
\description{
Applies variable names changes to \code{GADSdat} or \code{all_GADSdat} objects.
}
\details{
tbd.
}
\examples{
# Example data set
#to be done

}
