% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{collapseColumns}
\alias{collapseColumns}
\title{Collapse columns.}
\usage{
collapseColumns(lookUp, recodeVars, prioritize)
}
\arguments{
\item{recodeVars}{Character vector of variable names which should be recoded.}

\item{GADSdat}{A \code{GADSdat} object.}

\item{addCols}{Character vector of additional column names for recoding purposes.}

\item{sort_by}{By which column should the long format data.frame be sorted? If \code{NULL}, no sorting is performed.}
}
\value{
Returns a data frame that can be used for \code{\link{applyLookup}}.
}
\description{
Extract values form variables of \code{GADSdat} object and write to \code{xlsx} for manuel recoding.
}
\details{
If recoding of one or multiple variables is more complex, a lookup table can be created for later importing via \code{eatGADS}.
}
