% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{collapseColumns}
\alias{collapseColumns}
\title{Collapse columns.}
\usage{
collapseColumns(lookup, recodeVars, prioritize)
}
\arguments{
\item{lookup}{A lookup table \code{data.frame} as created via \code{\link{createLookup}}.}

\item{recodeVars}{Character vector of variable names which should be collapsed.}

\item{prioritize}{Which of the variables should be prioritized, if multiple values are available?}
}
\value{
Returns a data frame that can be used for \code{\link{applyLookup}}.
}
\description{
Extract values form variables of \code{GADSdat} object and write to \code{xlsx} for manuel recoding.
}
\details{
If recoding of one or multiple variables is more complex, a lookup table can be created for later importing via \code{eatGADS}.
}
