% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{collapseMC_Text}
\alias{collapseMC_Text}
\title{Recode MC variable based on text.}
\usage{
collapseMC_Text(GADSdat, mc_var, text_var, mc_code4text, suffix = "_r")
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{mc_var}{A single variable name of the multiple choice variable.}

\item{text_var}{A single variable name of the text variable.}

\item{mc_code4text}{The value in the MC variable that signals that information from the text variable should be used.}

\item{suffix}{Variable suffix for the newly created GADSdat.}
}
\value{
Returns a \code{GADSdat} containing the newly computed variable.
}
\description{
Use an additional text variable to recode an existing MC variable.
}
\details{
to be written
}
