% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGADS.R
\name{createGADS}
\alias{createGADS}
\title{Create an \code{eatGADS} data base.}
\usage{
createGADS(allList, pkList, fkList, filePath)
}
\arguments{
\item{allList}{An object created via \code{\link{mergeLabels}}.}

\item{pkList}{List of primary keys.}

\item{fkList}{List of foreign keys.}

\item{filePath}{Path to the db file to write (including name); has to end on '.db'.}
}
\value{
Creates a data base in the given path, returns \code{NULL}.
}
\description{
Creates a relational data base containing hierarchically stored data with meta information (e.g. value and variable labels).
}
\details{
Uses \code{\link[eatDB]{createDB}} from the \code{eatDB} package to create a relational data base. For details on how to define
keys see the documentation of \code{\link[eatDB]{createDB}}.
}
\examples{
# see createDB vignette

}
