% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_recode.R
\name{createLookup}
\alias{createLookup}
\title{Recode values.}
\usage{
createLookup(GADSdat, recodeVars, sort_by = NULL, addCols = c("value_new"))
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{recodeVars}{Character vector of variable names which should be recoded.}

\item{sort_by}{By which column should the long format data.frame be sorted? If \code{NULL}, no sorting is performed.}

\item{addCols}{Character vector of additional column names for recoding purposes.}
}
\value{
Returns a data frame in long format including all unique values of the variables in \code{recodeVars}.
}
\description{
Extract values form variables of \code{GADSdat} object and write to \code{xlsx} for manuel recoding.
}
\details{
If recoding of one or multiple variables is more complex, a lookup table can be created for later importing via \code{eatGADS}.
}
\examples{
#to be done

}
