% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVariable.R
\name{createVariable}
\alias{createVariable}
\title{Create a variable.}
\usage{
createVariable(GADSdat, varName)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{varName}{Name of the variable to be cloned.}
}
\value{
Returns a \code{GADSdat}.
}
\description{
Create an empty variable as part of a \code{GADSdat} object.
}
\examples{
# create a new variable
pisa_new <- createVariable(pisa, varName = "new_variable")

}
