% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyTheseVariables.R
\name{emptyTheseVariables}
\alias{emptyTheseVariables}
\title{Set variables to \code{NA}.}
\usage{
emptyTheseVariables(GADSdat, vars)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{vars}{Character vector of variable names which should be set to \code{NA}.}
}
\value{
Returns the recoded \code{GADSdat}.
}
\description{
Set all values within one or multiple variables to \code{NA}.
}
\examples{
# empty multiple variables
pisa2 <- emptyTheseVariables(pisa, vars = c("idstud", "idschool"))
}
