% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalGADS.R
\name{equalGADS}
\alias{equalGADS}
\title{Test if Two \code{GADSdat} Objects are (Nearly) Equal}
\usage{
equalGADS(target, current, metaExceptions = NULL)
}
\arguments{
\item{target}{A \code{GADSdat} object.}

\item{current}{A \code{GADSdat} object.}

\item{metaExceptions}{Should certain meta data columns be excluded from the comparison?}
}
\value{
Returns a list.
}
\description{
Run tests to check whether two \code{GADSdat} objects are (nearly) equal. Variable names, number of rows in the data,
meta data and data differences are checked and reported as a list output.
}
