% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tibble.R
\name{export_tibble}
\alias{export_tibble}
\title{Transform a \code{GADSdat} to a \code{tibble}}
\usage{
export_tibble(GADSdat)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}
}
\value{
Returns a \code{tibble}.
}
\description{
\code{haven}'s \code{\link[haven]{read_spss}} stores data together with meta data (e.g. value and variable labels) in a
\code{tibble} with attributes on variable level. This function transforms a \code{GADSdat} object to such a \code{tibble}.
}
\details{
This function is mainly intended for internal use.
}
\examples{
pisa_tbl <- export_tibble(pisa)

}
