% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{extractData}
\alias{extractData}
\title{Extract Data}
\usage{
extractData(GADSdat, convertMiss = TRUE, convertLabels = "character",
  dropPartialLabels = TRUE, convertVariables)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{convertMiss}{Should values labeled as missings be recoded to \code{NA}?}

\item{convertLabels}{If \code{"numeric"}, values remain as numerics. If \code{"factor"} or \code{"character"}, values are recoded to their labels. Corresponding variable type is applied.}

\item{dropPartialLabels}{Should partially complete labels be dropped? Often, this is the desired behaviour.}

\item{convertVariables}{Character vector of variables names, which labels should be applied to. If not specified, value labels are applied to all variables.}
}
\value{
Returns a data frame.
}
\description{
Extract data.frame from a \code{GADSdat} object for analyses in R.
}
\details{
A \code{GADSdat} object includes actual data and the corresponding meta data information. extractData extracts the data and applies relevant meta information (missing conversion, value labels), so the data can be used for analyses in R. Careful: If factors are extracted, the underlying integers will not equal the original underlying integers and \code{as.numeric} will probably yield undesired results.
}
\examples{
# Example data set
#to be done

}
