% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGADSdat.R
\name{extractGADSdat}
\alias{extractGADSdat}
\title{Extract single \code{GADSdat} from \code{all_GADSdat}}
\usage{
extractGADSdat(all_GADSdat, name)
}
\arguments{
\item{all_GADSdat}{\code{all_GADSdat} object}

\item{name}{A character vector with length 1 with the name of the \code{GADSdat}}
}
\value{
Returns an \code{GADSdat} object.
}
\description{
Function to extract a single \code{GADSdat} from an \code{all_GADSdat} object.
}
\details{
\code{GADSdat} objects can be merged into a single \code{all_GADSdat} object via \code{\link{mergeLabels}}. This function, performs the
reverse action, extracting a single \code{GADSdat} object.
}
\examples{
# see createGADS vignette

}
