% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{extractMeta}
\alias{extractMeta}
\title{Get Meta Data}
\usage{
extractMeta(GADSobject, vars = NULL)
}
\arguments{
\item{GADSobject}{A \code{GADSdat} object.}

\item{vars}{A character vector containing variable names. If \code{NULL} (default), all available metainformation is returned.}
}
\value{
Returns a long format data frame with meta information.
}
\description{
Exctrat meta data (e.g. variable and values labels) from an \code{eatGADS} object. This can be a \code{GADSdat}, an \code{all_GADSdat}, a labels \code{data.frame}, or the path to an existing data base.
}
\details{
Meta data is stored tidily in all GADS objects as a separate long format data frame. This information can be extracted for a single or multiple variables.
}
\examples{
\dontrun{
# Extract Meta data from data base
metaData <- extractMeta(GADSobject = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db", vars = "domain")

# Extract Meta data from loaded GADS
gads10 <- getGADS(vSelect = c("idstud", "wgt", "jkzone", "jkrep", "imp", "domain", "score"),
                 filePath = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")
metaData <- extractMeta(gads10, vars = "domain")
}

}
