% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{extractMeta}
\alias{extractMeta}
\title{Get Metainformation and Labels}
\usage{
extractMeta(GADS_object, vars)
}
\arguments{
\item{vars}{A character vector containing variable names.}

\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
Returns a long format data frame with meta information.
}
\description{
Exctrat metainformation, variable and values labels from GADSdat.
}
\details{
Metainformation is stored tidily in a GADSdat and can be extracted via extractMeta for a single or multiple variables.
}
\examples{
# Example data set
#to be done

}
