% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{extractMeta}
\alias{extractMeta}
\title{Get Metainformation and Labels}
\usage{
extractMeta(GADS_object, vars = NULL)
}
\arguments{
\item{GADS_object}{A \code{GADSdat} object.}

\item{vars}{A character vector containing variable names. If \code{NULL} (default), alle available metainformation is returned.}
}
\value{
Returns a long format data frame with meta information.
}
\description{
Exctrat metainformation, variable and values labels from an \code{eatGADS} object with meta information. This can be a \code{GADSdat}, an \code{all_GADSdat}, a labels \code{data.frame}, or the path to an existing data base.
}
\details{
Metainformation is stored tidily in a GADSdat and can be extracted via extractMeta for a single or multiple variables.
}
\examples{
# Example data set
#to be done

}
