% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_program_limit.R
\name{get_program_limit}
\alias{get_program_limit}
\title{Get program specific limits}
\usage{
get_program_limit(
  program = c("SPSS", "Stata"),
  component = c("all", "varNames", "varLabels", "valLabels", "stringvars", "nrows",
    "ncols"),
  version = NULL
)
}
\arguments{
\item{program}{Character vector. Name(s) of the program(s) whose limits should be returned.}

\item{component}{Optional single string. Which limits should be returned? If this is "all" or
not specified, all limits for the chosen \code{program}(s) are returned.}

\item{version}{Optional. Use a specific version of a program (see details).}
}
\value{
A list of two elements: \code{x} (numeric size of the limit) and
 \code{unit} ("char", "byte", or "generic"). If more than one \code{program} is specified and
 \code{component} is not "all", the most restrictive limit will be returned.
}
\description{
Get different limits for datasets specific to \code{SPSS} and \code{Stata}. This function
 primarily serves as a helper to the (sub)functions of the dataset compliance check family.
}
\details{
\code{SPSS} and \code{Stata} impose different limits to different aspects of their data sets:
\itemize{
 \item \code{varNames}: length of variable names
 \item \code{varLabels}: length of variable labels
 \item \code{valLabels}: length of value labels
 \item \code{stringvars}: length of strings in character variables
 \item \code{nrows}: number of observations
 \item \code{ncols}: number of variables
}
Run \code{get_program_limit()} to see all limits that are collected in this function.

\subsection{Software versions}{
Different versions of the software can entail different limits. With \code{SPSS}, there only
 differences are between older and newer versions. With \code{Stata}, however, limits differ
 between product versions [1]. By default, limits of \code{SPSS 30} and \code{Stata 19/SE}
 are used.

 Limits of different versions can be requested with the \code{version} argument.
 This overwrites \code{program}, but allows only for a single program and version. Currently,
 the following program versions are implemented: \code{c("Stata 19/BE", "Stata 19/MP")}
}
}
\examples{
# Get all limits that Stata imposes on its datasets
eatGADS::get_program_limit("Stata", "all")

# Get the specific limit on variable name lengths under SPSS
eatGADS::get_program_limit("SPSS", "varNames")
}
\references{
[1] \href{https://www.stata.com/products/comparison-of-limits/}{Stata: Comparison of limits}
}
\seealso{
Other dataset compliance checks: 
\code{\link{checkIntOverflow}()},
\code{\link{checkLabeledFractionals}()}
}
\concept{dataset compliance checks}
