% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{getGADS}
\alias{getGADS}
\title{Get data from GADS data base.}
\usage{
getGADS(vSelect = NULL, filePath)
}
\arguments{
\item{vSelect}{Variables}

\item{filePath}{Path of the existing db file.}
}
\value{
Returns a GADSdat object.
}
\description{
Extracts variables from a GADS data base. Only the specified variables are extracted. Note that this selection determines the format of the \code{data.frame} that is extracted.
}
\details{
See \code{\link[eatDB]{createDB}} and \code{\link[eatDB]{dbPull}} for further explanation of the query and merging processes.
}
\examples{
\dontrun{
gads10 <- getGADS(vSelect = c("idstud", "wgt", "jkzone", "jkrep", "imp", "domain", "score"),
                 filePath = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")
# View Meta Data
metaData <- extractMeta(gads10)
# Extract Data for Analysis
dat <- extractData(gads10)
}

}
