% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{getGADS_fast}
\alias{getGADS_fast}
\title{Get data from GADS data base fast from server directory.}
\usage{
getGADS_fast(vSelect = NULL, filePath, tempPath = tempdir())
}
\arguments{
\item{vSelect}{Variables}

\item{filePath}{Path of the existing db file.}

\item{tempPath}{Local directory in which the dataBase can temporarily be stored. Using the default is recommended.}
}
\value{
Returns a GADSdat object.
}
\description{
Extracts variables from a GADS data base. Only the specified variables are extracted. Note that this selection determines the format of the \code{data.frame} that is extracted. Uses a local temporary directory to speed up loading the GADS from a server.
}
\details{
A random temporary directory is used for caching the data base and should be removed, when the computer is restarted. See \code{\link{createDB}} and \code{\link{dbPull}} for further explanation of the query and merging processes.
}
\examples{
\dontrun{
gads10 <- getGADS_fast(vSelect = c("idstud", "wgt", "jkzone", "jkrep", "imp", "domain", "score"),
                      filePath = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")
# View Meta Data
metaData <- extractMeta(gads10)
# Extract Data for Analysis
dat <- extractData(gads10)
}

}
