% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProgramLimit.R
\name{getProgramLimit}
\alias{getProgramLimit}
\title{Get program specific limits}
\usage{
getProgramLimit(
  program = c("SPSS", "Stata", "Stata 19/BE", "Stata 19/MP"),
  component = c("varNames", "varLabels", "valLabels", "stringvars", "nrows", "ncols")
)
}
\arguments{
\item{program}{Character vector of the programs/program version that should be considered.}

\item{component}{Single string. Which limits should be returned?}
}
\value{
A list of two elements: \code{value} (numeric size of the limit) and
 \code{unit} ("char", "byte", or "generic").
}
\description{
Get the (most restrictive) \link[=program_limits]{limits} that \code{SPSS} and/or \code{Stata}
 imposes on a specific aspect of a dataset.
}
\details{
For more details about program specific limits as well as a full list, see \link{program_limits}.
 In \code{program}, \code{"SPSS"} implies \code{SPSS 30}, and \code{"Stata"} implies
 \code{Stata 19/SE}, as these are the most relevant version among the ones implemented here.
 If more than one program/version name is given in \code{program}, the most restrictive limit
 will be returned.
}
\examples{
# Show all implemented limits
program_limits

# Get the specific limit on variable name lengths under SPSS
getProgramLimit("SPSS", "varNames")

# Get the variable name length limit a dataset has to adhere to to be compatible with
#  both SPSS and Stata 19/SE
getProgramLimit(c("Stata", "SPSS"), "varNames")
}
