% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_trend.R
\name{getTrendGADS}
\alias{getTrendGADS}
\title{Get data for trend reports.}
\usage{
getTrendGADS(filePath1, filePath2, lePath = NULL, vSelect = NULL,
  years, fast = TRUE, tempPath = tempdir())
}
\arguments{
\item{filePath1}{Path of the first GADS db file.}

\item{filePath2}{Path of the second GADS db file.}

\item{lePath}{Path of the linking error db file. If NULL, no linking errors are added to the data.}

\item{vSelect}{Variables from both GADS to be selected (as character vector).}

\item{years}{A numeric vector of length 2. The first elements corresponds to filePath1, the second element to filePath2.}

\item{fast}{Should \code{\link{getGADS_fast}} be used for data loading instead of \code{\link{getGADS}}? Using the default is heavily recommended.}

\item{tempPath}{The directory, in which both GADS will be temporarily stored. Using the default is heavily recommended.}
}
\value{
Returns a GADSdat object.
}
\description{
Extracts variables from two GADS data bases and a linking error data base. Data can then be extracted from the \code{GADSdat} object via \code{\link{extractData}}. For extracting meta data from a db file or a \code{GADSdat} object see \code{\link{extractMeta}}. To speed up the data loading, \code{\link{getGADS_fast}} is used pre default.
}
\details{
This function extracts data from two GADS data bases and a linking error data base. All data bases have to be created via \code{\link{createGADS}}. The two GADS are joined via \code{rbind} and a variable \code{year} is added, corresponding to the argument \code{years}. If \code{lePath} is specified, linking errors are also extracted and then merged to the GADS data. Make sure to also extract the key variables necessary for merging the linking errors (the domain variable for all linking errors, additionally the competence level variable for linking errors for competence levels). The \code{GADSdat} object can then further be used via \code{\link{extractData}}. See \code{\link[eatDB]{createDB}} and \code{\link[eatDB]{dbPull}} for further explanation of the querying and merging processes.
}
\examples{
\dontrun{
trend_gads <- getTrendGADS(filePath1 = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db",
                          filePath2 = "t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2015.db",
                          lePath = "t:/_R_Tutorials/R_Workshops/04_eatPakete/les_2010_2015.db",
                          years = c(2010, 2015),
                          vSelect = c("idstud", "wgt", "jkzone", "jkrep", "imp", "domain", "score"))

# Extract Data
dat <- extractData(trend_gads)

# Extract Meta Data
extractMeta(trend_gads)

}

}
