% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertVariable.R
\name{insertVariable}
\alias{insertVariable}
\title{Reorder a single variable in a \code{GADSdat}.}
\usage{
insertVariable(GADSdat, var, after = NULL)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{var}{Character string of the variable name which should be sorted.}

\item{after}{Character string of the variable name after which \code{var} should be inserted. If \code{NULL}, \code{var} is inserted at the beginning of the
\code{GADSdat}.}
}
\description{
Deprecated. Please use \code{\link{relocateVariable}} instead.
}
