% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLabels.R
\name{mergeLabels}
\alias{mergeLabels}
\title{Prepare data and metadata}
\usage{
mergeLabels(...)
}
\arguments{
\item{...}{\code{GADSdat} objects, as named arguments in the correct merge order.}
}
\value{
Returns an \code{all_GADSdat} object, which consists of list with a list of all data frames \code{"datList"} and a single data frame containing all meta data information \code{"allLabels"}.
}
\description{
Transform multiple \code{GADSdat} objects into a list ready for data base creation.
}
\details{
The function \code{\link{createGADS}} takes multiple \code{GADSdat} objects as input. The function preserves the ordering
in which the objects are supplied, which is then used for the merging order in \code{\link{createGADS}}. Additionally,
the separate lists of meta information for each \code{GADSdat} are merged and a data frame unique identifier is added.
}
\examples{
# see createGADS vignette

}
