% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_data_handling.R
\name{miss2NA}
\alias{miss2NA}
\title{Recode Missings to \code{NA}}
\usage{
miss2NA(GADSdat)
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}
}
\value{
Returns the data frame with \code{NA} instead of missing codes.
}
\description{
Recode Missings to \code{NA} according to missing labels in label data frame.
}
\details{
Missings are imported as their values via \code{\link{import_spss}}. Using the value labels in the labels data frame, \code{miss2NA} recodes these missings codes to \code{NA}.
}
\examples{
# Example data set
#to be done

}
