% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_create_GADS.R
\name{namesGADS}
\alias{namesGADS}
\title{Variables names of a GADS data base.}
\usage{
namesGADS(filePath)
}
\arguments{
\item{filePath}{Path of an existing GADS data base.}
}
\value{
Returns a named list of variable names per data table.
}
\description{
Returns a list of all variable names included in the GADS data base.
}
\details{
Extracts names of all variables included in the relational data base, structured as list with the individual data tables as list elements.
}
\examples{
# not run:
# dbNames(filePath = "exampleDB.db")

}
