% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_eatDB_UI.R
\name{namesGADS}
\alias{namesGADS}
\title{Variables names of a GADS data base.}
\usage{
namesGADS(filePath)
}
\arguments{
\item{filePath}{Path of an existing GADS data base.}
}
\value{
Returns a named list of variable names per data table.
}
\description{
Returns a list of all variable names included in the GADS data base.
}
\details{
Extracts names of all variables included in the relational data base, structured as list with the individual data tables as list elements.
}
\examples{
\dontrun{
varNames <- namesGADS("t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")
}

}
