% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namesGADS.R
\name{namesGADS}
\alias{namesGADS}
\title{Variables names of a GADS data base.}
\usage{
namesGADS(GADS)
}
\arguments{
\item{GADS}{Path of an existing \code{eatGADS} data base.}
}
\value{
Returns a named list of variable names per data table.
}
\description{
Returns a list of all variable names included in the GADS data base.
}
\details{
Extracts names of all variables included in the relational data base, structured as a list with the individual data tables as list elements.
}
\examples{
\dontrun{
varNames <- namesGADS("t:/_R_Tutorials/R_Workshops/04_eatPakete/minigads_2010.db")

# all variable names sorted by data table
varNames

# variables in a specific data table
varNames$allDat

}

}
