% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa.R
\docType{data}
\name{pisa}
\alias{pisa}
\title{PISA Plus Example Data}
\format{
A data.frame with 500 rows and 133 variables, including:
\describe{
  \item{idstud}{Person ID variable}
  \item{idschool}{School ID variable}
  \item{schtype}{School type}
  \item{...}{}
}
}
\source{
Research Data Center at the Institute for Educational Quality Improvement (2020). Programme for International Student Assessment - Plus 2012, 2013 (PISA Plus 2012-2013) - Campus File (Version 1) [Data set]. Berlin: Institute for Educational Quality Improvement. \doi{https://doi.org/10.5159/IQB_PISA_Plus_2012-13_CF_v1}
}
\usage{
pisa
}
\description{
A small example data set from the German PISA Plus campus files as distributed by the \code{Forschungsdatenzentrum, IQB}.
}
\keyword{datasets}
