% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_limits.R
\docType{data}
\name{program_limits}
\alias{program_limits}
\title{Program specific limits to dataset components}
\format{
A data.frame listing relevant limits (see details) imposed to datasets by
 \code{SPSS} and \code{Stata}.
}
\usage{
program_limits
}
\description{
Different programs impose different limits to different components of their datasets.
 Additionally, limits may vary between software versions. This primarily applies to
 \code{Stata}'s product tiers, but also to (very) old \code{SPSS} versions. \code{eatGADS}
 offers a number of check functions - chiefly \code{\link{check4SPSS}} and
 \code{\link{check4Stata}} as wrappers - to ensure a \code{GADSdat} complies with these limits,
 and can be exported into an \link[=write_spss]{SPSS} or \link[=write_stata]{Stata} file.
 Use \code{\link{getProgramLimit}} for a more convenient interface for obtaining specific limits.
}
\details{
While datasets have several components and characteristics, the following were deemed the most
 important and their limits implemented in this package's checks:
\itemize{
 \item \code{varNames}: length of variable names
 \item \code{varLabels}: length of variable labels
 \item \code{valLabels}: length of value labels
 \item \code{stringvars}: length of strings in character variables
 \item \code{nrows}: number of observations
 \item \code{ncols}: number of variables
}

While \code{SPSS} has only one set of limits (disregarding legacy limits for older versions),
 \code{Stata} employs different limits for different product versions [1]. Within this package,
 \code{"SPSS"} always implies \code{SPSS 30}, and \code{"Stata"} implies \code{Stata 19/SE}.
 Limits of \code{Stata 19/BE} and \code{Stata 19/MP} are implemented as additional options.
 However, no additional versions of \code{SPSS} have been implemented yet.
}
\references{
[1] \href{https://www.stata.com/products/comparison-of-limits/}{Stata: Comparison of limits}
}
\keyword{datasets}
