% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_metadata_handling.R
\name{recodeGADS}
\alias{recodeGADS}
\title{Recode a variable.}
\usage{
recodeGADS(GADSdat, varName, oldValues, newValues, newValueLabels = NULL)
}
\arguments{
\item{GADSdat}{GADSdat object imported via eatGADS.}

\item{varName}{Name of the variable to be recoded.}

\item{oldValues}{Vector containing the old values.}

\item{newValues}{Vector containing the new values, in identical order as oldNames.}

\item{newValueLabels}{Named vector containing new value labels for the new values. All new values have to get lables. Slightly experimental implementation.}
}
\value{
Returns the GADSdat object with changed variable names.
}
\description{
Recode a variable as part of a \code{GADSdat} or \code{all_GADSdat} object.
}
\details{
Applied to a \code{GADSdat} or \code{all_GADSdat} object, this function is a wrapper of \code{\link{getChangeMeta}} and \code{\link{applyChangeMeta}}
}
\examples{
# Example data set
#to be done

}
