% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeString2NA.R
\name{recodeString2NA}
\alias{recodeString2NA}
\title{Recode a string to \code{NA}.}
\usage{
recodeString2NA(GADSdat, recodeVars = namesGADS(GADSdat), string = "")
}
\arguments{
\item{GADSdat}{A \code{GADSdat} object.}

\item{recodeVars}{Character vector of variable names which should be recoded.}

\item{string}{Which string should be recoded to \code{NA}?}
}
\value{
Returns the recoded \code{GADSdat}.
}
\description{
Deprecated, use \code{\link{recode2NA}} instead..
}
