% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeValLabels.R
\name{removeValLabels}
\alias{removeValLabels}
\title{Remove value labels.}
\usage{
removeValLabels(GADSdat, varName, value)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via \code{eatGADS}.}

\item{varName}{Character string of a variable name.}

\item{value}{Numeric values.}
}
\value{
Returns the \code{GADSdat} object with changed meta data.
}
\description{
Remove value labels of a variable as part of a \code{GADSdat} or \code{all_GADSdat} object.
}
\examples{
# Example data set
#to be done

}
