% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_metadata_handling.R
\name{reuseMeta}
\alias{reuseMeta}
\title{Use meta data for a variable from another GADSdat.}
\usage{
reuseMeta(GADSdat, varName, other_GADSdat, other_varName = NULL,
  dropMissingLabels = FALSE)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via eatGADS.}

\item{varName}{Name of the variable that should get the new meta data.}

\item{other_GADSdat}{\code{GADSdat} object imported via eatGADS including the desired meta information.}

\item{other_varName}{Name of the variable that should get the new meta data in the \code{other_GADSdat}.}

\item{dropMissingLabels}{Should value labels for missings be transfered to? Relevant for imputed data sets for which missing labels are usually dropped.}
}
\value{
Returns the original object with updated meta data.
}
\description{
Transfer meta information from one GADSdat to another.
}
\details{
tbd.
}
\examples{
# Example data set
#to be done

}
