% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02b_metadata_handling.R
\name{reuseMeta}
\alias{reuseMeta}
\title{Use meta data for a variable from another GADSdat.}
\usage{
reuseMeta(GADSdat, varName, other_GADSdat, other_varName = NULL,
  missingLabels = NULL)
}
\arguments{
\item{GADSdat}{\code{GADSdat} object imported via eatGADS.}

\item{varName}{Name of the variable that should get the new meta data.}

\item{other_GADSdat}{\code{GADSdat} object imported via eatGADS including the desired meta information. Can also be a GADS db or an \code{all_GADSdat} object.}

\item{other_varName}{Name of the variable that should get the new meta data in the \code{other_GADSdat}.}

\item{missingLabels}{How should meta data for missing values be treated? If NULL, missings are transfered as all other labels. If "drop", missing labels are dropped (useful for imputed data). If "leave", missing labels remain untouched.}
}
\value{
Returns the original object with updated meta data.
}
\description{
Transfer meta information from one GADSdat to another.
}
\details{
tbd.
}
\examples{
# Example data set
#to be done

}
