% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_write_spss.R
\name{write_spss}
\alias{write_spss}
\title{Write data frame to sav file}
\usage{
write_spss(GADSdat, filePath)
}
\arguments{
\item{GADSdat}{Data frame.}

\item{filePath}{Path of sav file to write.}
}
\value{
Writes sav-file, returns nothing.
}
\description{
Preliminary Version! Function to write data frame with labels for values and variables to SPSS-file.
}
\details{
...
}
\examples{
# tbd

}
