% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_write_spss.R
\name{writeDB_SPSS}
\alias{writeDB_SPSS}
\title{Write data frame to sav file}
\usage{
writeDB_SPSS(df, label_df, filePath)
}
\arguments{
\item{df}{Data frame.}

\item{filePath}{Path of the existing db file.}

\item{filePath}{Path of sav file to write.}
}
\value{
Writes sav-file, returns nothing.
}
\description{
Function to write data frame with labels for values and variables to SPSS-file.
}
\details{
...
}
\examples{
# # See vignette.

}
